pyproject.toml
[tool.poetry]
name = "gitctx"
version = "0.1.0"
description = ""
authors = ["kvanzuijlen <developer@debijenkorf.nl>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
GitPython = "^3.1.30"

[tool.poetry.scripts]
gitctx = 'gitctx.__main__:cli'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
